chore: MovieLabel을 파일로 분리
